+++ include/hb/player_physics.h
@@ -4,17 +4,17 @@
 #include <hb/collision_flags.h>
 
 struct hb_player_physics {
-	float                                 gravity[2];
-	float                                     radius;
-	float                                   inv_mass;
-	float                                     b_coef;
-	float                                    damping;
+	double                                gravity[2];
+	double                                    radius;
+	double                                  inv_mass;
+	double                                    b_coef;
+	double                                   damping;
 	enum hb_collision_flags                  c_group;
-	float                               acceleration;
-	float                       kicking_acceleration;
-	float                            kicking_damping;
-	float                              kick_strength;
-	float                                   kickback;
+	double                              acceleration;
+	double                      kicking_acceleration;
+	double                           kicking_damping;
+	double                             kick_strength;
+	double                                  kickback;
 };
 
 #endif
